court_scraper/requirements.txt
flask
requests
beautifulsoup4
fpdf
mysql-connector-python
